+++ AdaptivePELE/simulation/simulationrunner.py
@@ -80,6 +80,9 @@ class SimulationParameters:
         self.mpiParameters = None
         self.numberEquilibrationStructures = 10
         self.reportName = ""
+        self.equilibrationBoxRadius = 2
+        self.equilibrationRotationRange = 0.05
+        self.equilibrationTranslationRange = 0.5
         # parameters needed for MD simulations and their defaults
         self.timeStep = 2
         self.ligandCharge = 0
@@ -429,6 +432,19 @@ class PeleSimulation(SimulationRunner):
         PDBinitial.initialise(initialStruct, resname=resname, chain=reschain, resnum=resnum)
         return repr(PDBinitial.getCOM())
 
+    def generatePELECommand(self, runningControlFile):
+        """
+        Generate the command to run PELE using the given parameters
+
+        :param runningControlFile: Path of the control file to run
+        :type runningControlFile: str
+        """
+
+        if self.parameters.srun:
+            return ["srun", "-n", str(self.parameters.processors)] + self.parameters.srunParameters + [self.parameters.executable, runningControlFile]
+        else:
+            return ["mpirun", "-np", str(self.parameters.processors)] + self.parameters.mpiParameters + [self.parameters.executable, runningControlFile]
+
     def runEquilibrationPELE(self, runningControlFile):
         """
         Run a short PELE equilibration simulation
@@ -438,10 +454,7 @@ class PeleSimulation(SimulationRunner):
         """
 
         self.createSymbolicLinks()
-        if self.parameters.srun:
-            toRun = ["srun", "-n", str(self.parameters.processors)] + self.parameters.srunParameters + [self.parameters.executable, runningControlFile]
-        else:
-            toRun = ["mpirun", "-np", str(self.parameters.processors)] + self.parameters.mpiParameters + [self.parameters.executable, runningControlFile]
+        toRun = self.generatePELECommand(runningControlFile)
 
         utilities.print_unbuffered(" ".join(toRun))
         startTime = time.time()
@@ -479,10 +492,8 @@ class PeleSimulation(SimulationRunner):
         self.prepareControlFile(epoch, outputPathConstants, ControlFileDictionary)
         self.createSymbolicLinks()
         runningControlFile = outputPathConstants.tmpControlFilename % epoch
-        if self.parameters.srun:
-            toRun = ["srun", "-n", str(self.parameters.processors)] + self.parameters.srunParameters + [self.parameters.executable, runningControlFile]
-        else:
-            toRun = ["mpirun", "-np", str(self.parameters.processors)] + self.parameters.mpiParameters + [self.parameters.executable, runningControlFile]
+        toRun = self.generatePELECommand(runningControlFile)
+
         utilities.print_unbuffered(" ".join(toRun))
         startTime = time.time()
         if self.parameters.time:
@@ -524,12 +535,11 @@ class PeleSimulation(SimulationRunner):
             return peleControlFileDict
 
         # Set small rotations and translations
-        peleControlFileDict["commands"][0]["Perturbation"]["parameters"]["translationRange"] = 0.5
-        peleControlFileDict["commands"][0]["Perturbation"]["parameters"]["rotationScalingFactor"] = 0.05
+        peleControlFileDict["commands"][0]["Perturbation"]["parameters"]["translationRange"] = self.parameters.equilibrationTranslationRange
+        peleControlFileDict["commands"][0]["Perturbation"]["parameters"]["rotationScalingFactor"] = self.parameters.equilibrationRotationRange
         if "Box" in peleControlFileDict["commands"][0]["Perturbation"]:
-            # Set box_radius to 2
             peleControlFileDict["commands"][0]["Perturbation"]["Box"]["fixedCenter"] = "$BOX_CENTER"
-            peleControlFileDict["commands"][0]["Perturbation"]["Box"]["radius"] = 2
+            peleControlFileDict["commands"][0]["Perturbation"]["Box"]["radius"] = "$BOX_RADIUS"
         # Ensure random tags exists in metrics
         metricsBlock = peleControlFileDict["commands"][0]["PeleTasks"][0]["metrics"]
         nMetrics = len(metricsBlock)
@@ -543,7 +553,7 @@ class PeleSimulation(SimulationRunner):
         # Add equilibration dynamical changes
         changes = {
             "doThesechanges": {
-                "Perturbation::parameters": {"rotationScalingFactor": 0.050}
+                "Perturbation::parameters": {"rotationScalingFactor": 0.05}
             },
             "ifAnyIsTrue": ["rand >= .5"],
             "otherwise": {
@@ -619,7 +629,9 @@ class PeleSimulation(SimulationRunner):
 
             :returns: list --  List with initial structures
         """
-        if not any([x is not None for x in (resname, resnum, reschain)]):
+        # check that we have some way of identifyin the ligand
+        ligandInputs = [resname != "", resnum != 0, reschain != ""]
+        if not any(ligandInputs):
             raise utilities.RequiredParameterMissingException("Ligand information not specified in clustering block!!!")
         newInitialStructures = []
         newStructure = []
@@ -640,7 +652,7 @@ class PeleSimulation(SimulationRunner):
             equilibrationPeleDict["OUTPUT_PATH"] = equilibrationOutput
             equilibrationPeleDict["COMPLEXES"] = initialStructureString
             equilibrationPeleDict["BOX_CENTER"] = self.selectInitialBoxCenter(structure, resname, reschain, resnum)
-            equilibrationPeleDict["BOX_RADIUS"] = 2
+            equilibrationPeleDict["BOX_RADIUS"] = self.parameters.equilibrationBoxRadius
             equilibrationPeleDict["REPORT_NAME"] = reportFilename
             equilibrationPeleDict["TRAJECTORY_NAME"] = trajName
             for name in ["BOX_CENTER", "BOX_RADIUS"]:
@@ -702,7 +714,7 @@ class PeleSimulation(SimulationRunner):
             raise utilities.UnsatisfiedDependencyException("No installation of scikit-learn found. Please, install scikit-learn or select a different equilibrationMode.")
         energyColumn = 3
         # detect number of trajectories available
-        nTrajs = len(glob.glob(trajWildcard.rsplit("_", 1)[0]+"*"))+1
+        nTrajs = len(utilities.getReportList(trajWildcard.rsplit("_", 1)[0]+"*"))
         data = []
         for i in range(1, nTrajs):
             report = utilities.loadtxtfile(reportWildcard % i)
@@ -713,7 +725,7 @@ class PeleSimulation(SimulationRunner):
                 com = conformation.getCOM()
                 data.append([line[energyColumn], i, nSnap]+com)
         if not data:
-            raise utilities.UnspecifiedPELECrashException("Some happened with PELE and no trajectories were written!!")
+            raise utilities.UnspecifiedPELECrashException("Some error happened with PELE and no trajectories were written!!")
         data = np.array(data)
         data = data[data[:, 0].argsort()]
         nPoints = max(self.parameters.numberEquilibrationStructures, data.shape[0]//4)
@@ -1347,11 +1359,14 @@ class TestSimulation(SimulationRunner):
         self.prepareControlFile(epoch, outputPathConstants, ControlFileDictionary)
         if not self.copied:
             tmp_sync = os.path.join(outputPathConstants.tmpFolder, os.path.split(processManager.syncFolder)[1])
+            tmp_topology = os.path.join(outputPathConstants.tmpFolder, os.path.split(topologies.path)[1])
             shutil.copytree(processManager.syncFolder, tmp_sync)
+            shutil.copytree(topologies.path, tmp_topology)
             if os.path.exists(self.parameters.destination):
                 shutil.rmtree(self.parameters.destination)
             shutil.copytree(self.parameters.origin, self.parameters.destination)
             shutil.copytree(tmp_sync, processManager.syncFolder)
+            shutil.copytree(tmp_topology, topologies.path)
             self.copied = True
 
     def makeWorkingControlFile(self, workingControlFilename, dictionary, inputTemplate=None):
@@ -1486,6 +1501,9 @@ class RunnerBuilder:
             params.equilibrationMode = paramsBlock.get(blockNames.SimulationParams.equilibrationMode, blockNames.SimulationParams.equilibrationSelect)
             params.equilibrationLength = paramsBlock.get(blockNames.SimulationParams.equilibrationLength)
             params.numberEquilibrationStructures = paramsBlock.get(blockNames.SimulationParams.numberEquilibrationStructures, 10)
+            params.equilibrationBoxRadius = paramsBlock.get(blockNames.SimulationParams.equilibrationBoxRadius, 2)
+            params.equilibrationRotationRange = paramsBlock.get(blockNames.SimulationParams.equilibrationRotationRange, 0.05)
+            params.equilibrationTranslationRange = paramsBlock.get(blockNames.SimulationParams.equilibrationTranslationRange, 0.5)
             params.srun = paramsBlock.get(blockNames.SimulationParams.srun, False)
             params.trajsPerReplica = params.processors
             params.numReplicas = 1